azure_integration/requirements.txt
asttokens==2.0.5
azure-devops==5.0.0b5
backcall==0.2.0
cachetools==5.0.0
certifi==2021.10.8
charset-normalizer==2.0.12
decorator==5.1.1
executing==0.8.3
google-api-core==2.7.2
google-auth==2.6.4
google-cloud-bigquery==3.0.1
google-cloud-bigquery-storage==2.13.1
google-cloud-core==2.3.0
google-crc32c==1.3.0
google-resumable-media==2.3.2
googleapis-common-protos==1.56.0
grpcio==1.44.0
grpcio-status==1.44.0
idna==3.3
ipdb==0.13.9
ipython==8.2.0
isodate==0.6.1
jedi==0.18.1
matplotlib-inline==0.1.3
msrest==0.6.21
numpy==1.22.3
oauthlib==3.2.0
packaging==21.3
pandas==1.4.2
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.29
proto-plus==1.20.3
protobuf==3.20.0
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==7.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.11.2
pyparsing==3.0.8
python-dateutil==2.8.2
pytz==2022.1
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
six==1.16.0
stack-data==0.2.0
toml==0.10.2
traitlets==5.1.1
urllib3==1.26.9
wcwidth==0.2.5
